requirements.txt
requests>=2.32.0
APScheduler>=3.10.4
python-telegram-bot>=21.0
scikit-learn>=1.5.0
xgboost>=2.1.0
